+++ composeApp/src/commonMain/kotlin/dev/jketterer/leaflog/presentation/ui/screens/settings/teatype/TeaTypeListScreen.kt
@@ -202,7 +202,7 @@ private fun TeaTypeCard(
                 val details = buildList {
                     teaType.defaultTemperatureCelsius?.let { celsius ->
                         val displayTemp = UnitConverter.celsiusToDisplayTemperature(
-                            celsius,
+                            celsius.toDouble(),
                             userPreferences.temperatureUnit,
                         )
                         add("$displayTemp${userPreferences.temperatureUnit.symbol}")
